reorganize and rename files, functions
